refactor: code review
Log:
