+++ src/modules/primary-output/outputmanagement.h
@@ -16,7 +16,7 @@ WAYLIB_SERVER_USE_NAMESPACE
 
 class PrimaryOutputV1
     : public QObject
-    , public Waylib::Server::WServerInterface
+    , public WAYLIB_SERVER_NAMESPACE::WServerInterface
 {
     Q_OBJECT
     Q_PROPERTY(const char *primaryOutput READ primaryOutput WRITE setPrimaryOutput NOTIFY primaryOutputChanged)
