+++ src/modules/shortcut/shortcutmanager.h
@@ -21,7 +21,7 @@ WAYLIB_SERVER_USE_NAMESPACE
 
 class ShortcutV1
     : public QObject
-    , public Waylib::Server::WServerInterface
+    , public WAYLIB_SERVER_NAMESPACE::WServerInterface
 {
     Q_OBJECT
