+++ src/modules/window-management/windowmanagement.h
@@ -12,7 +12,7 @@ WAYLIB_SERVER_USE_NAMESPACE
 
 class WindowManagementV1
     : public QObject
-    , public Waylib::Server::WServerInterface
+    , public WAYLIB_SERVER_NAMESPACE::WServerInterface
 {
     Q_OBJECT
     Q_PROPERTY(DesktopState desktopState READ desktopState WRITE setDesktopState NOTIFY desktopStateChanged)
