+++ src/qml/PrimaryOutput.qml
@@ -89,6 +89,7 @@ OutputItem {
         workspace: Helper.workspace.current
         anchors.fill: parent
         fillMode: Image.PreserveAspectCrop
+        // TODO: qt6.8 use other property
         asynchronous: false
         clip: true
