+++ src/treeland/treeland.cpp
@@ -6,160 +6,40 @@
 #include "cmdline.h"
 #include "compositor1adaptor.h"
 #include "helper.h"
-#include "shellhandler.h"
 
 #include <Messages.h>
 #include <SignalHandler.h>
 #include <SocketWriter.h>
 
-#include <WCursor>
-#include <WSeat>
-#include <WSurface>
-#include <wordexp.h>
-#include <woutputrenderwindow.h>
-#include <wrenderhelper.h>
 #include <wsocket.h>
 #include <wxwayland.h>
 
-#include <qwbackend.h>
-#include <qwcompositor.h>
-#include <qwdatacontrolv1.h>
-#include <qwdisplay.h>
-#include <qwlogging.h>
-#include <qwoutput.h>
-#include <qwxwayland.h>
-
-#include <DLog>
-
 #include <QDebug>
-#include <QEventLoop>
-#include <QGuiApplication>
+#include <QLocalSocket>
 #include <QLoggingCategory>
-#include <QQmlApplicationEngine>
-#include <QQmlContext>
-#include <QQmlEngine>
-#include <QQuickItem>
-#include <QQuickStyle>
-#include <QQuickView>
-#include <QTimer>
-#include <QtLogging>
 
 #include <pwd.h>
 #include <sys/socket.h>
 #include <unistd.h>
 
-Q_LOGGING_CATEGORY(treelandMain, "treeland.dbus", QtDebugMsg);
+Q_LOGGING_CATEGORY(dbus, "treeland.dbus", QtDebugMsg);
 
 using namespace DDM;
 DCORE_USE_NAMESPACE;
-
-QString unescape(const QString &str) noexcept
-{
-    QString unescapedStr;
-    for (qsizetype i = 0; i < str.size(); ++i) {
-        auto c = str.at(i);
-        if (c != '\\') {
-            unescapedStr.append(c);
-            continue;
-        }
-
-        switch (str.at(i + 1).toLatin1()) {
-        default:
-            unescapedStr.append(c);
-            break;
-        case 'n':
-            unescapedStr.append('\n');
-            ++i;
-            break;
-        case 't':
-            unescapedStr.append('\t');
-            ++i;
-            break;
-        case 'r':
-            unescapedStr.append('\r');
-            ++i;
-            break;
-        case '\\':
-            unescapedStr.append('\\');
-            ++i;
-            break;
-        case ';':
-            unescapedStr.append(';');
-            ++i;
-            break;
-        case 's': {
-            unescapedStr.append(R"(\ )");
-            ++i;
-        } break;
-        }
-    }
-
-    return unescapedStr;
-}
-
-std::optional<QStringList> unescapeExecArgs(const QString &str) noexcept
-{
-    auto unescapedStr = unescape(str);
-    if (unescapedStr.isEmpty()) {
-        qCWarning(treelandMain) << "unescape Exec failed.";
-        return std::nullopt;
-    }
-
-    auto deleter = [](wordexp_t *word) {
-        wordfree(word);
-        delete word;
-    };
-    std::unique_ptr<wordexp_t, decltype(deleter)> words{ new (std::nothrow)
-                                                             wordexp_t{ 0, nullptr, 0 },
-                                                         deleter };
-
-    if (auto ret = wordexp(unescapedStr.toLocal8Bit(), words.get(), WRDE_SHOWERR); ret != 0) {
-        QString errMessage;
-        switch (ret) {
-        case WRDE_BADCHAR:
-            errMessage = "BADCHAR";
-            break;
-        case WRDE_BADVAL:
-            errMessage = "BADVAL";
-            break;
-        case WRDE_CMDSUB:
-            errMessage = "CMDSUB";
-            break;
-        case WRDE_NOSPACE:
-            errMessage = "NOSPACE";
-            break;
-        case WRDE_SYNTAX:
-            errMessage = "SYNTAX";
-            break;
-        default:
-            errMessage = "unknown";
-        }
-        qCWarning(treelandMain) << "wordexp error: " << errMessage;
-        return std::nullopt;
-    }
-
-    QStringList execList;
-    for (std::size_t i = 0; i < words->we_wordc; ++i) {
-        execList.emplace_back(words->we_wordv[i]);
-    }
-
-    return execList;
-}
 
 namespace TreeLand {
 TreeLand::TreeLand(Helper *helper)
     : QObject()
     , m_helper(helper)
+    , m_socket(new QLocalSocket(this))
 {
-    m_socket = new QLocalSocket(this);
-
     connect(m_socket, &QLocalSocket::connected, this, &TreeLand::connected);
     connect(m_socket, &QLocalSocket::disconnected, this, &TreeLand::disconnected);
     connect(m_socket, &QLocalSocket::readyRead, this, &TreeLand::readyRead);
     connect(m_socket, &QLocalSocket::errorOccurred, this, &TreeLand::error);
 
     if (CmdLine::ref().socket().has_value()) {
-        new DDM::SignalHandler;
+        new DDM::SignalHandler(this);
         Q_ASSERT(m_helper);
         auto connectToServer = [this] {
             m_socket->connectToServer(CmdLine::ref().socket().value());
@@ -177,15 +57,14 @@ TreeLand::TreeLand(Helper *helper)
 
     if (CmdLine::ref().run().has_value()) {
         auto exec = [runCmd = CmdLine::ref().run().value(), this] {
-            qCInfo(treelandMain) << "run cmd:" << runCmd;
-            if (auto cmdline = unescapeExecArgs(runCmd); cmdline) {
+            qCInfo(dbus) << "run cmd:" << runCmd;
+            if (auto cmdline = CmdLine::ref().unescapeExecArgs(runCmd); cmdline) {
                 auto cmdArgs = cmdline.value();
 
                 auto envs = QProcessEnvironment::systemEnvironment();
                 envs.insert("WAYLAND_DISPLAY", m_helper->defaultWaylandSocket()->fullServerName());
                 envs.insert("DISPLAY", m_helper->defaultXWaylandSocket()->displayName());
                 envs.insert("DDE_CURRENT_COMPOSITOR", "TreeLand");
-                envs.insert("QT_IM_MODULE", "fcitx");
 
                 QProcess process;
                 process.setProgram(cmdArgs.constFirst());
@@ -234,7 +113,7 @@ bool TreeLand::debugMode() const
 void TreeLand::connected()
 {
     // log connection
-    qCDebug(treelandMain) << "Connected to the daemon.";
+    qCDebug(dbus) << "Connected to the daemon.";
 
     // send connected message
     DDM::SocketWriter(m_socket) << quint32(DDM::GreeterMessages::Connect);
@@ -248,11 +127,11 @@ void TreeLand::retranslate() noexcept
 void TreeLand::disconnected()
 {
     // log disconnection
-    qCDebug(treelandMain) << "Disconnected from the daemon.";
+    qCDebug(dbus) << "Disconnected from the daemon.";
 
     Q_EMIT socketDisconnected();
 
-    qCDebug(treelandMain) << "Display Manager is closed socket connect, quiting treeland.";
+    qCDebug(dbus) << "Display Manager is closed socket connect, quiting treeland.";
     qApp->exit();
 }
 
@@ -274,7 +153,7 @@ void TreeLand::readyRead()
         switch (DDM::DaemonMessages(message)) {
         case DDM::DaemonMessages::Capabilities: {
             // log message
-            qCDebug(treelandMain) << "Message received from daemon: Capabilities";
+            qCDebug(dbus) << "Message received from daemon: Capabilities";
         } break;
         case DDM::DaemonMessages::LoginSucceeded:
         case DDM::DaemonMessages::UserActivateMessage: {
@@ -350,17 +229,15 @@ QString TreeLand::XWaylandName()
     QProcess *process = new QProcess(this);
     connect(process, &QProcess::finished, [process, m, conn, user, display] {
         if (process->exitCode() != 0) {
-            qCWarning(treelandMain)
-                << "xhost command failed with exit code" << process->exitCode()
-                << process->readAllStandardOutput() << process->readAllStandardError();
+            qCWarning(dbus) << "xhost command failed with exit code" << process->exitCode()
+                            << process->readAllStandardOutput() << process->readAllStandardError();
             auto reply =
                 m.createErrorReply(QDBusError::InternalError, "Failed to set xhost permissions");
             conn.send(reply);
         } else {
-            qCDebug(treelandMain) << process->exitCode() << " " << process->readAllStandardOutput()
-                                  << process->readAllStandardError();
-            qCDebug(treelandMain)
-                << QString("user %1 got xwayland display %2.").arg(user).arg(display);
+            qCDebug(dbus) << process->exitCode() << " " << process->readAllStandardOutput()
+                          << process->readAllStandardError();
+            qCDebug(dbus) << QString("user %1 got xwayland display %2.").arg(user).arg(display);
             auto reply = m.createReply(display);
             conn.send(reply);
         }
