+++ src/treeland/wallpapermanager.cpp
@@ -22,7 +22,7 @@ WallpaperManager *WallpaperManager::instance()
     return instance;
 }
 
-void WallpaperManager::add(WallpaperImage *proxy, Waylib::Server::WOutputItem *outputItem)
+void WallpaperManager::add(WallpaperImage *proxy, WAYLIB_SERVER_NAMESPACE::WOutputItem *outputItem)
 {
     Q_ASSERT(m_proxys.find(outputItem) == m_proxys.end());
     m_proxys[outputItem] = proxy;
@@ -37,7 +37,7 @@ void WallpaperManager::remove(WallpaperImage *proxy)
     m_proxys.remove(m_proxys.key(proxy));
 }
 
-void WallpaperManager::remove(Waylib::Server::WOutputItem *outputItem)
+void WallpaperManager::remove(WAYLIB_SERVER_NAMESPACE::WOutputItem *outputItem)
 {
     m_proxys.remove(outputItem);
 }
@@ -66,7 +66,7 @@ bool WallpaperManager::isSelfLocked(const WallpaperController *controller) const
     return m_proxyLockList.contains(controller);
 }
 
-WallpaperImage *WallpaperManager::get(Waylib::Server::WOutputItem *outputItem) const
+WallpaperImage *WallpaperManager::get(WAYLIB_SERVER_NAMESPACE::WOutputItem *outputItem) const
 {
     if (!outputItem) {
         return nullptr;
@@ -75,7 +75,7 @@ WallpaperImage *WallpaperManager::get(Waylib::Server::WOutputItem *outputItem) c
     return get(outputItem->output());
 }
 
-WallpaperImage *WallpaperManager::get(Waylib::Server::WOutput *output) const
+WallpaperImage *WallpaperManager::get(WAYLIB_SERVER_NAMESPACE::WOutput *output) const
 {
     for (auto *proxy : m_proxys.keys()) {
         if (proxy->output() == output) {
